refactor(feature-manager): delete remaining unused code (#2223)
* refactor(feature-manager): delete FeatureManager class

* refactor(feature-manager): delete MissingFeatureError

* refactor(feature-manager): delete FeatureManager types

* refactor(feature-manager): rename config dir to features

* test: fix remaining test
