--- src/app/config/feature-manager/index.ts
@@ -1,7 +1,0 @@
-import FeatureManager from './util/FeatureManager.class'
-
-export * from './types'
-
-const featureManager = new FeatureManager()
-
-export default featureManager
