+++ src/app/config/features/intranet.config.ts
