+++ src/app/modules/verified-content/verified-content.types.ts
@@ -2,7 +2,6 @@ import { Result } from 'neverthrow'
 
 import { VerifiedKeys } from '../../../shared/util/verified-content'
 import { AuthType } from '../../../types'
-import { MissingFeatureError } from '../core/core.errors'
 
 import { MalformedVerifiedContentError } from './verified-content.errors'
 
@@ -27,10 +26,6 @@ export type SpVerifiedContent = {
 }
 
 export type VerifiedContentResult<T> = Result<T, MalformedVerifiedContentError>
-export type FactoryVerifiedContentResult<T, E> = Result<
-  T,
-  E | MissingFeatureError
->
 
 export type EncryptVerificationContentParams = {
   verifiedContent: CpVerifiedContent | SpVerifiedContent
